Rearrange project files
