--- app/src/main/java/com/example/trw/sampleroompersistence/callback/SendDataCallback.java
@@ -1,10 +1,0 @@
-package com.example.trw.sampleroompersistence.callback;
-
-/**
- * Created by TRW on 5/10/2560.
- */
-
-public interface SendDataCallback {
-
-    void findPlayerIdCallback(int playerId, boolean status);
-}
