--- app/src/main/java/com/example/trw/sampleroompersistence/callback/SendStatusCallback.java
@@ -1,10 +1,0 @@
-package com.example.trw.sampleroompersistence.callback;
-
-/**
- * Created by TRW on 2/10/2560.
- */
-
-public interface SendStatusCallback {
-
-    void checkStatusCallback(boolean status);
-}
