+++ app/src/main/java/com/example/trw/sampleroompersistence/db/callback/SendDataCallback.java
@@ -1,0 +1,10 @@
+package com.example.trw.sampleroompersistence.db.callback;
+
+/**
+ * Created by TRW on 5/10/2560.
+ */
+
+public interface SendDataCallback {
+
+    void findPlayerIdCallback(int playerId, boolean status);
+}
