+++ app/src/main/java/com/example/trw/sampleroompersistence/db/callback/SendStatusCallback.java
@@ -1,0 +1,10 @@
+package com.example.trw.sampleroompersistence.db.callback;
+
+/**
+ * Created by TRW on 2/10/2560.
+ */
+
+public interface SendStatusCallback {
+
+    void checkStatusCallback(boolean status);
+}
