--- app/src/main/java/com/example/trw/sampleroompersistence/recyclerview/holder/BaseViewHolder.java
@@ -1,15 +1,0 @@
-package com.example.trw.sampleroompersistence.recyclerview.holder;
-
-import android.support.v7.widget.RecyclerView;
-import android.view.View;
-
-/**
- * Created by TRW on 2/10/2560.
- */
-
-public class BaseViewHolder extends RecyclerView.ViewHolder {
-
-    public BaseViewHolder(View itemView) {
-        super(itemView);
-    }
-}
