--- app/src/main/java/com/example/trw/sampleroompersistence/recyclerview/item/BaseItem.java
@@ -1,18 +1,0 @@
-package com.example.trw.sampleroompersistence.recyclerview.item;
-
-/**
- * Created by TRW on 2/10/2560.
- */
-
-public class BaseItem {
-
-    private int type;
-
-    public BaseItem(int type) {
-        this.type = type;
-    }
-
-    public int getType() {
-        return type;
-    }
-}
