+++ app/src/main/java/com/example/trw/sampleroompersistence/ui/adapter/ViewType.java
@@ -1,0 +1,12 @@
+package com.example.trw.sampleroompersistence.ui.adapter;
+
+/**
+ * Created by TRW on 15/8/2560.
+ */
+
+public class ViewType {
+
+    public static final int TYPE_LIST_PLAYER_AWARD = 1;
+    public static final int TYPE_LIST_PLAYER_CONTRACT = 2;
+
+}
