+++ app/src/main/java/com/example/trw/sampleroompersistence/ui/adapter/holder/BaseViewHolder.java
@@ -1,0 +1,15 @@
+package com.example.trw.sampleroompersistence.ui.adapter.holder;
+
+import android.support.v7.widget.RecyclerView;
+import android.view.View;
+
+/**
+ * Created by TRW on 2/10/2560.
+ */
+
+public class BaseViewHolder extends RecyclerView.ViewHolder {
+
+    public BaseViewHolder(View itemView) {
+        super(itemView);
+    }
+}
