+++ app/src/main/java/com/example/trw/sampleroompersistence/ui/adapter/item/BaseItem.java
@@ -1,0 +1,18 @@
+package com.example.trw.sampleroompersistence.ui.adapter.item;
+
+/**
+ * Created by TRW on 2/10/2560.
+ */
+
+public class BaseItem {
+
+    private int type;
+
+    public BaseItem(int type) {
+        this.type = type;
+    }
+
+    public int getType() {
+        return type;
+    }
+}
