+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.example.trw.sampleroompersistence.MainActivity">
+    tools:context="com.example.trw.sampleroompersistence.ui.MainActivity">
 
     <android.support.design.widget.TabLayout
         android:id="@+id/tab_layout_menu"
