+++ app/src/main/res/layout/fragment_award.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.example.trw.sampleroompersistence.fragment.AwardFragment">
+    tools:context="com.example.trw.sampleroompersistence.ui.award.AwardFragment">
 
     <LinearLayout
         android:layout_width="wrap_content"
