+++ app/src/main/res/layout/fragment_home.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.trw.sampleroompersistence.fragment.HomeFragment">
+    tools:context="com.example.trw.sampleroompersistence.ui.home.HomeFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
