+++ app/src/main/res/layout/fragment_manage.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:scrollbars="none"
-    tools:context="com.example.trw.sampleroompersistence.fragment.ManageFragment">
+    tools:context="com.example.trw.sampleroompersistence.ui.manage.ManageFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
